Flowers & Trees sway in the wind (#271)
## About The Pull Request

This PR animates flower bushes and trees to sway as if they were being
blown by wind. Russians did it first. Also this PR title would make a
great band name

## Why It's Good For The Game

yea

testing eviden: had to keep it short so that i dont have to upload it to
youtube and do a whole process



https://github.com/user-attachments/assets/60c51533-2b8a-4350-a70e-23a408c0b8bd



## Changelog



:cl:

image: flowers & trees now sway in the wind


/:cl:

---------

Co-authored-by: FalloutFalcon <86381784+FalloutFalcon@users.noreply.github.com>